requirements.txt
albumentations==1.3.0
opencv-python>=4.5.0
pybboxes>=0.1.0
PyYAML>=6.0